Switch to boto3 for S3 object retrieval (#23)
The S3 Proxy service is going to migrate away from PaaS and be hosted on
an EC2 Instance (or a cluster of them). The desired auth method for S3
access will change from IAM key-based access to IAM Instance Role
permissions. This requires that the S3 object retrieval part of the
proxy functionality be handled via the S3 API, rather than signed
headers over HTTP as the proxy was doing until now.
It was decided to use the `boto3` library to handle the S3 interaction,
and this would allow the proxy to be run in the current setup or under
the new credential system.

> Note that this update causes breaking changes to the way the proxy
must be run, in either the existing PaaS or new system - some ENV vars
have changed name (for compatibility with boto's naming scheme) or have
been added or made optional as a result of moving from path- to
key-based object retrieval.